Check if All the Integers in a Range Are Covered
Runtime: 1ms 53.31%
Memory: 40.2mb 94.81%

I would iterate through the array of integer arrays and set corresponding indices in the boolean array to true if it was covered by the ranges. Then, i would iterate through the range from the values of left and right and check if any index in the corresponding boolean array was false and returned false if so. True if otherwise.